WIP init command rewrite
