+++ cmd/new.go
@@ -1,0 +1,36 @@
+package cmd
+
+import (
+	"github.com/spf13/afero"
+	"github.com/spf13/cobra"
+)
+
+// newCmd represents the new command
+var newCmd = &cobra.Command{
+	Use:   "new",
+	Short: "Create a new change file",
+	Long: `Creates a new change file.
+Change files are processed when preparing a new release and will create a changelog for the new version.
+Each version is combined together for the overall project changelog.`,
+	RunE: runNew,
+}
+
+func init() {
+	rootCmd.AddCommand(newCmd)
+
+	// Here you will define your flags and configuration settings.
+
+	// Cobra supports Persistent Flags which will work for this command
+	// and all subcommands, e.g.:
+	// newCmd.PersistentFlags().String("foo", "", "A help for foo")
+
+	// Cobra supports local flags which will only run when this command
+	// is called directly, e.g.:
+	// newCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
+
+func runNew(cmd *cobra.Command, args []string) error {
+	fs := afero.NewOsFs()
+	config := project.LoadConfig(fs)
+	return nil
+}
