+++ go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/onsi/gomega v1.10.3
 	github.com/spf13/afero v1.1.2
 	github.com/spf13/cobra v1.1.1
+	gopkg.in/yaml.v2 v2.3.0
 )
