+++ project/change.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"time"
+)
+
+type Change struct {
+	kind string
+	body string
+	time time.Time
+}
